fnal task

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
 
 void print_number(int n)
 {
-	unsigned int m, d, count;
+	short  int m, d, count;
 
 	if (n < 0)
 	{
